update: Handle vendor onboarding migration from version 5.
