+++ includes/Upgrade/Upgrades.php
@@ -45,7 +45,6 @@ class Upgrades {
         '3.7.19' => Upgrades\V_3_7_19::class,
         '3.13.0' => Upgrades\V_3_13_0::class,
         '3.14.0' => Upgrades\V_3_14_0::class,
-        '4.2.2'  => Upgrades\V_4_2_2::class,
         '5.0.0'  => Upgrades\V_5_0_0::class,
     ];
